@@ -6,29 +6,13 @@ from ....worker.credentials import CloudUserCredentials
 
 
 class AzureUserCredentials(CloudUserCredentials):
-    def __init__(self, data: Dict[str, bytes]):
+    def __init__(self, data: Dict[str, str]):
         self._data = data
         self._credentials = json.loads(base64.b64decode(data['key.json']).decode())
 
     @property
-    def secret_name(self) -> str:
-        return 'azure-credentials'
-
-    @property
-    def secret_data(self) -> Dict[str, bytes]:
-        return self._data
-
-    @property
-    def file_name(self) -> str:
-        return 'key.json'
-
-    @property
     def cloud_env_name(self) -> str:
         return 'AZURE_APPLICATION_CREDENTIALS'
-
-    @property
-    def hail_env_name(self) -> str:
-        return 'HAIL_AZURE_CREDENTIAL_FILE'
 
     @property
     def username(self):
@@ -40,7 +24,7 @@ class AzureUserCredentials(CloudUserCredentials):
 
     @property
     def mount_path(self):
-        return f'/{self.secret_name}/{self.file_name}'
+        return '/azure-credentials/key.json'
 
     def cloudfuse_credentials(self, fuse_config: dict) -> str:
         # https://github.com/Azure/azure-storage-fuse
