@@ -5,28 +5,12 @@ from ....worker.credentials import CloudUserCredentials
 
 
 class GCPUserCredentials(CloudUserCredentials):
-    def __init__(self, data: Dict[str, bytes]):
+    def __init__(self, data: Dict[str, str]):
         self._data = data
-
-    @property
-    def secret_name(self) -> str:
-        return 'gsa-key'
-
-    @property
-    def secret_data(self) -> Dict[str, bytes]:
-        return self._data
-
-    @property
-    def file_name(self) -> str:
-        return 'key.json'
 
     @property
     def cloud_env_name(self) -> str:
         return 'GOOGLE_APPLICATION_CREDENTIALS'
-
-    @property
-    def hail_env_name(self) -> str:
-        return 'HAIL_GSA_KEY_FILE'
 
     @property
     def username(self):
@@ -34,11 +18,11 @@ class GCPUserCredentials(CloudUserCredentials):
 
     @property
     def password(self) -> str:
-        return base64.b64decode(self.secret_data['key.json']).decode()
+        return base64.b64decode(self._data['key.json']).decode()
 
     @property
     def mount_path(self):
-        return f'/{self.secret_name}/{self.file_name}'
+        return '/gsa-key/key.json'
 
     def cloudfuse_credentials(self, fuse_config: dict) -> str:  # pylint: disable=unused-argument
         return self.password
