@@ -47,7 +47,7 @@ class GCPWorkerAPI(CloudWorkerAPI):
     def get_compute_client(self) -> aiogoogle.GoogleComputeClient:
         return self._compute_client
 
-    def user_credentials(self, credentials: Dict[str, bytes]) -> GCPUserCredentials:
+    def user_credentials(self, credentials: Dict[str, str]) -> GCPUserCredentials:
         return GCPUserCredentials(credentials)
 
     async def worker_access_token(self, session: httpx.ClientSession) -> Dict[str, str]:
