@@ -1,39 +1,26 @@
 import abc
-from typing import Dict
 
 
 class CloudUserCredentials(abc.ABC):
     @property
-    def secret_name(self) -> str:
-        raise NotImplementedError
-
-    @property
-    def secret_data(self) -> Dict[str, bytes]:
-        raise NotImplementedError
-
-    @property
-    def file_name(self) -> str:
-        raise NotImplementedError
-
-    @property
+    @abc.abstractmethod
     def cloud_env_name(self) -> str:
         raise NotImplementedError
 
     @property
-    def hail_env_name(self) -> str:
-        raise NotImplementedError
-
-    @property
+    @abc.abstractmethod
     def username(self) -> str:
         raise NotImplementedError
 
     @property
+    @abc.abstractmethod
     def password(self) -> str:
         raise NotImplementedError
 
     @property
+    @abc.abstractmethod
     def mount_path(self):
-        return f'/{self.secret_name}/{self.file_name}'
+        raise NotImplementedError
 
     @abc.abstractmethod
     def cloudfuse_credentials(self, fuse_config: dict) -> str:
