@@ -93,7 +93,7 @@ class BatchPoolExecutor:
         Backend used to execute the jobs. Must be a :class:`.ServiceBackend`.
     image:
         The name of a Docker image used for each submitted job. The image must
-        include Python 3.7 or later and must have the ``dill`` Python package
+        include Python 3.8 or later and must have the ``dill`` Python package
         installed. If you intend to use ``numpy``, ensure that OpenBLAS is also
         installed. If unspecified, an image with a matching Python verison and
         ``numpy``, ``scipy``, and ``sklearn`` installed is used.
@@ -138,9 +138,9 @@ class BatchPoolExecutor:
         self._shutdown = False
         version = sys.version_info
         if image is None:
-            if version.major != 3 or version.minor not in (7, 8, 9, 10):
+            if version.major != 3 or version.minor not in (8, 9, 10):
                 raise ValueError(
-                    f'You must specify an image if you are using a Python version other than 3.7, 3.8, 3.9 or 3.10 (you are using {version})')
+                    f'You must specify an image if you are using a Python version other than 3.8, 3.9 or 3.10 (you are using {version})')
             self.image = f'hailgenetics/python-dill:{version.major}.{version.minor}-slim'
         else:
             self.image = image
