@@ -865,7 +865,7 @@ class PythonJob(Job):
 
         # Create a batch object with a default Python image
 
-        b = Batch(default_python_image='hailgenetics/python-dill:3.7-slim')
+        b = Batch(default_python_image='hailgenetics/python-dill:3.8-slim')
 
         def multiply(x, y):
             return x * y
@@ -923,11 +923,11 @@ class PythonJob(Job):
         Examples
         --------
 
-        Set the job's docker image to `hailgenetics/python-dill:3.7-slim`:
+        Set the job's docker image to `hailgenetics/python-dill:3.8-slim`:
 
         >>> b = Batch()
         >>> j = b.new_python_job()
-        >>> (j.image('hailgenetics/python-dill:3.7-slim')
+        >>> (j.image('hailgenetics/python-dill:3.8-slim')
         ...   .call(print, 'hello'))
         >>> b.run()  # doctest: +SKIP
 
